move relevant tests from codegen to extractor
